Shadow Walls
Has the use of the rotate180 function that allows the shadow walls to flip robot direction and dictate movement. With further use of the getTheta() function and random variables we can get the robots to bounce of the walls randomly away from the direction they came in from. Currently they just flip 180.